refactor(open-mpm): split init/ and telegram/ oversized files (#366)
