+++ crates/open-mpm/src/telegram/handlers.rs
@@ -1,0 +1,468 @@
+//! Slash-command and plain-text message handlers for the Telegram gateway.
+//!
+//! Why: The dptree dispatch routes updates to these two functions; keeping
+//! them separate from the pairing/format/lifecycle concerns keeps each file
+//! focused and under the 500-line cap.
+//! What: The `Command` enum (teloxide `BotCommands` derive), `handle_command`
+//! (slash dispatch with the pairing gate), and `handle_message` (plain-text +
+//! `/switch` persona routing into ctrl).
+//! Test: Exercised indirectly; the pure helpers they call are unit-tested in
+//! `telegram::tests`.
+
+use std::path::PathBuf;
+use std::sync::Arc;
+use std::time::Instant;
+
+use teloxide::prelude::*;
+use teloxide::types::{ChatAction, ParseMode};
+use teloxide::utils::command::BotCommands;
+use teloxide::utils::html;
+use tracing::warn;
+
+use super::format::{markdown_to_html_safe, send_long_html};
+use super::pairing::{
+    PAIRING_CODE_TTL, PairOutcome, PairedChats, PendingPairs, SENTINEL_PAIRING_CHAT_ID,
+    save_paired_chats, verify_pair_attempt,
+};
+use super::{ChatSession, SessionMap, home_persona_exists};
+use crate::ctrl::{self, ConversationTurn};
+
+/// Slash commands exposed to Telegram users.
+///
+/// Why: teloxide's `BotCommands` derive auto-generates parsing + a help string
+/// from the `description` attributes. Routing is done in the dptree handler.
+/// What: `/start`, `/help`, `/connect <path>`, `/clear`, `/status`.
+/// Test: Compile-time check via `BotCommands::descriptions()`.
+#[derive(BotCommands, Clone, Debug)]
+#[command(
+    rename_rule = "lowercase",
+    description = "open-mpm Telegram bot commands:"
+)]
+pub(super) enum Command {
+    #[command(description = "Welcome message and pairing code")]
+    Start,
+    #[command(description = "Authorize this chat with a pairing code: /pair <code>")]
+    Pair(String),
+    #[command(description = "List available commands")]
+    Help,
+    #[command(description = "Set the project path for this chat: /connect <path>")]
+    Connect(String),
+    #[command(description = "Reset conversation history for this chat")]
+    Clear,
+    #[command(description = "Show current project path and LLM backend")]
+    Status,
+}
+
+/// Dispatch a parsed slash command.
+// Why: teloxide's dptree dispatch passes injected state as positional
+// arguments; this function reflects the dispatch signature exactly so a
+// wrapper struct would just have to be unpacked at every call site.
+#[allow(clippy::too_many_arguments)]
+pub(super) async fn handle_command(
+    bot: Bot,
+    msg: Message,
+    cmd: Command,
+    sessions: SessionMap,
+    project_path: Arc<PathBuf>,
+    paired: PairedChats,
+    paired_state_path: Arc<PathBuf>,
+    pending: PendingPairs,
+) -> ResponseResult<()> {
+    let chat_id = msg.chat.id;
+
+    // #334: Gate every command except /start and /pair behind the pairing
+    // check. Unpaired chats get a uniform "send /start" prompt instead of
+    // partial information about the bot.
+    let is_unauthenticated_cmd = matches!(cmd, Command::Start | Command::Pair(_));
+    if !is_unauthenticated_cmd {
+        let is_paired = paired.read().await.contains_key(&chat_id);
+        if !is_paired {
+            bot.send_message(chat_id, "🔒 Not paired. Send /start to begin.")
+                .parse_mode(ParseMode::Html)
+                .await?;
+            return Ok(());
+        }
+    }
+
+    match cmd {
+        Command::Start => {
+            // #334: The pairing code is generated by the REPL (trusted
+            // terminal), NOT here. We only tell the user how to obtain it.
+            // This prevents Telegram-side attackers from self-authorizing
+            // even if they own the bot — they'd also need shell access to
+            // the host running the REPL.
+            tracing::info!(
+                chat_id = chat_id.0,
+                "Telegram /start received; instructing user to pair via REPL"
+            );
+            let _ = &pending; // pending is populated by the REPL, not here.
+            let text = concat!(
+                "🔐 <b>Pairing required</b>\n\n",
+                "To link this Telegram chat, go to your open-mpm REPL and run:\n\n",
+                "  <code>/telegram pair</code>\n\n",
+                "Then send the code here:  <code>/pair &lt;code&gt;</code>\n\n",
+                "(Codes expire in 5 minutes.)"
+            );
+            bot.send_message(chat_id, text)
+                .parse_mode(ParseMode::Html)
+                .await?;
+        }
+        Command::Pair(code_arg) => {
+            let provided = code_arg.trim().to_string();
+            if provided.is_empty() {
+                bot.send_message(chat_id, "Usage: <code>/pair &lt;code&gt;</code>")
+                    .parse_mode(ParseMode::Html)
+                    .await?;
+                return Ok(());
+            }
+            let now = Instant::now();
+            // #334: Look up first by per-chat key (legacy / future), then
+            // fall back to the REPL-issued sentinel entry. The sentinel is
+            // the standard path now.
+            let (outcome, matched_key) = {
+                let map = pending.lock().await;
+                let chat_outcome =
+                    verify_pair_attempt(map.get(&chat_id.0), &provided, now, PAIRING_CODE_TTL);
+                if matches!(
+                    chat_outcome,
+                    PairOutcome::Success | PairOutcome::Mismatch | PairOutcome::Expired
+                ) {
+                    (chat_outcome, chat_id.0)
+                } else {
+                    let sentinel_outcome = verify_pair_attempt(
+                        map.get(&SENTINEL_PAIRING_CHAT_ID),
+                        &provided,
+                        now,
+                        PAIRING_CODE_TTL,
+                    );
+                    (sentinel_outcome, SENTINEL_PAIRING_CHAT_ID)
+                }
+            };
+            match outcome {
+                PairOutcome::NoPending => {
+                    bot.send_message(chat_id, "No pending pairing. Send /start first.")
+                        .await?;
+                }
+                PairOutcome::Expired => {
+                    pending.lock().await.remove(&matched_key);
+                    bot.send_message(chat_id, "Code expired. Send /start to get a new code.")
+                        .await?;
+                }
+                PairOutcome::Mismatch => {
+                    bot.send_message(chat_id, "Invalid code.").await?;
+                }
+                PairOutcome::Success => {
+                    pending.lock().await.remove(&matched_key);
+                    paired.write().await.insert(chat_id, now);
+                    // #467: Persist the new pairing so it survives restart.
+                    // We deliberately log-and-continue on IO errors: losing
+                    // persistence is recoverable on the next successful save,
+                    // and we don't want a disk-full to block the user reply.
+                    if let Err(e) = save_paired_chats(&paired, &paired_state_path).await {
+                        warn!(
+                            chat_id = chat_id.0,
+                            error = %e,
+                            "failed to persist paired-chats state"
+                        );
+                    }
+                    tracing::info!(
+                        chat_id = chat_id.0,
+                        matched_via = matched_key,
+                        "Chat paired successfully"
+                    );
+                    bot.send_message(
+                        chat_id,
+                        "✅ <b>Paired successfully.</b> You can now send messages.",
+                    )
+                    .parse_mode(ParseMode::Html)
+                    .await?;
+                }
+            }
+        }
+        Command::Help => {
+            let text = format!(
+                "<b>Commands</b>\n\n<pre>{}</pre>",
+                html::escape(&Command::descriptions().to_string())
+            );
+            bot.send_message(chat_id, text)
+                .parse_mode(ParseMode::Html)
+                .await?;
+        }
+        Command::Connect(path_arg) => {
+            let trimmed = path_arg.trim();
+            if trimmed.is_empty() {
+                bot.send_message(chat_id, "Usage: <code>/connect &lt;path&gt;</code>")
+                    .parse_mode(ParseMode::Html)
+                    .await?;
+                return Ok(());
+            }
+            let new_path = PathBuf::from(trimmed);
+            if !new_path.is_dir() {
+                bot.send_message(
+                    chat_id,
+                    format!(
+                        "Path does not exist or is not a directory: <code>{}</code>",
+                        html::escape(trimmed)
+                    ),
+                )
+                .parse_mode(ParseMode::Html)
+                .await?;
+                return Ok(());
+            }
+            {
+                let mut map = sessions.lock().await;
+                let entry = map
+                    .entry(chat_id)
+                    .or_insert_with(|| ChatSession::new((*project_path).clone()));
+                entry.project_path = new_path.clone();
+                // /connect intentionally does NOT clear history — the user
+                // may want to reference earlier turns when changing project.
+            }
+            bot.send_message(
+                chat_id,
+                format!(
+                    "Connected to <code>{}</code>",
+                    html::escape(&new_path.display().to_string())
+                ),
+            )
+            .parse_mode(ParseMode::Html)
+            .await?;
+        }
+        Command::Clear => {
+            let mut map = sessions.lock().await;
+            if let Some(session) = map.get_mut(&chat_id) {
+                session.history.clear();
+            }
+            bot.send_message(chat_id, "Conversation history cleared.")
+                .await?;
+        }
+        Command::Status => {
+            let map = sessions.lock().await;
+            let path = map
+                .get(&chat_id)
+                .map(|s| s.project_path.clone())
+                .unwrap_or_else(|| (*project_path).clone());
+            let history_len = map.get(&chat_id).map(|s| s.history.len()).unwrap_or(0);
+            drop(map);
+
+            // #295: pass `None` here — the telegram status command has no
+            // single agent context. Passing None means claude-code will only
+            // surface if some downstream actually opts in (which they won't
+            // through this status path), and falls through to OpenRouter /
+            // Anthropic-direct when those env vars are set.
+            let llm_label = crate::llm::credentials::pick_credentials(None)
+                .map(|c| c.label())
+                .unwrap_or("none");
+
+            let text = format!(
+                "<b>Status</b>\n\n\
+                Project: <code>{}</code>\n\
+                Turns:   {}\n\
+                LLM:     <code>{}</code>",
+                html::escape(&path.display().to_string()),
+                history_len,
+                html::escape(llm_label)
+            );
+            bot.send_message(chat_id, text)
+                .parse_mode(ParseMode::Html)
+                .await?;
+        }
+    }
+    Ok(())
+}
+
+/// Forward a plain-text message to ctrl and reply with the result.
+pub(super) async fn handle_message(
+    bot: Bot,
+    msg: Message,
+    sessions: SessionMap,
+    project_path: Arc<PathBuf>,
+    paired: PairedChats,
+) -> ResponseResult<()> {
+    let chat_id = msg.chat.id;
+    let user_msg_id = msg.id;
+    let text = match msg.text() {
+        Some(t) => t.to_string(),
+        None => return Ok(()),
+    };
+
+    // #334: Gate the LLM dispatch behind the pairing check. We do this before
+    // showing the typing indicator so the user gets immediate feedback.
+    if !paired.read().await.contains_key(&chat_id) {
+        bot.send_message(chat_id, "🔒 Not paired. Send /start to begin.")
+            .parse_mode(ParseMode::Html)
+            .await?;
+        return Ok(());
+    }
+
+    // #457: Intercept `/switch <persona>` BEFORE ctrl dispatch. ctrl's
+    // text path passes the message verbatim to the LLM — it has no slash
+    // command awareness. We must handle persona switching ourselves and
+    // store the choice on the session so subsequent turns route through
+    // `run_pm_task_with_persona`.
+    if text.starts_with("/switch") {
+        return handle_switch(&bot, chat_id, &sessions, &project_path, &text).await;
+    }
+
+    // Show "typing…" indicator while we wait on the LLM. Best-effort: if it
+    // fails we still try the dispatch.
+    let _ = bot.send_chat_action(chat_id, ChatAction::Typing).await;
+
+    // Snapshot the project + history + active persona under the lock so we
+    // don't hold it across the multi-second LLM call.
+    let (path, history_snapshot, active_persona) = {
+        let mut map = sessions.lock().await;
+        let entry = map
+            .entry(chat_id)
+            .or_insert_with(|| ChatSession::new((*project_path).clone()));
+        (
+            entry.project_path.clone(),
+            entry.history.clone(),
+            entry.active_persona.clone(),
+        )
+    };
+
+    // #457: Route to the persona-specific runner if `/switch <persona>` is
+    // active on this chat. Otherwise the default ctrl agent handles it.
+    //
+    // #468: The default (no /switch) path used to call
+    // `run_pm_task_with_history`, which internally resolves the agent via
+    // `resolve_agent_config` — and that resolver prefers
+    // `{project}/.open-mpm/agents/pm.toml` over ctrl.toml. When the harness
+    // runs inside the open-mpm repo (or any project shipping a pm.toml),
+    // Telegram chats silently loaded the heavy sonnet PM prompt + full
+    // delegation toolset instead of the lightweight ctrl persona that
+    // ctrl.toml configures (e.g. `model = "ollama/qwen3:30b"`). The REPL
+    // already routes its conversational ctrl chat through
+    // `run_pm_task_with_persona("ctrl", …)` (see src/repl/dispatch.rs and
+    // `resolve_ctrl_agent_config`), which loads ctrl.toml directly. We
+    // mirror that pattern here so Telegram and REPL agree on which agent
+    // config wins by default.
+    let persona_name = active_persona.as_deref().unwrap_or("ctrl");
+    let result = ctrl::run_pm_task_with_persona(
+        &path,
+        persona_name,
+        &text,
+        &history_snapshot,
+        None,
+        ctrl::SessionOverrides::default(),
+    )
+    .await;
+
+    let response_text = match result {
+        Ok(reply) => {
+            // Persist the new turn. Do this before formatting so a parse
+            // error in the response doesn't drop the LLM round-trip.
+            let mut map = sessions.lock().await;
+            let entry = map
+                .entry(chat_id)
+                .or_insert_with(|| ChatSession::new((*project_path).clone()));
+            entry.history.push(ConversationTurn {
+                user: text.clone(),
+                assistant: reply.clone(),
+            });
+            drop(map);
+            markdown_to_html_safe(&reply)
+        }
+        Err(e) => {
+            warn!(chat_id = %chat_id.0, error = %e, "ctrl dispatch failed");
+            // #466: Surface the actual error to the user instead of a fixed
+            // "LLM backend not configured" string. The previous message was
+            // misleading: any failure (network, parse error, upstream 5xx)
+            // was reported as a credential problem. Including the real error
+            // makes debugging dramatically easier for operators.
+            format!("⚠️ Request failed: {e}")
+        }
+    };
+
+    send_long_html(&bot, chat_id, user_msg_id, &response_text).await;
+    Ok(())
+}
+
+/// Handle a `/switch <persona>` plain-text message.
+///
+/// Why: Persona switching is the only slash-like command we intercept in the
+/// plain-text path; extracting it keeps `handle_message` readable.
+/// What: Reports the current persona when no arg is given, validates the
+/// requested persona resolves (project-local then `~/.open-mpm/agents/`), and
+/// stores it on the session (or clears it for `ctrl`/`default`).
+/// Test: Exercised indirectly via the `/switch` flow.
+async fn handle_switch(
+    bot: &Bot,
+    chat_id: teloxide::types::ChatId,
+    sessions: &SessionMap,
+    project_path: &Arc<PathBuf>,
+    text: &str,
+) -> ResponseResult<()> {
+    let arg = text.trim_start_matches("/switch").trim().to_string();
+    if arg.is_empty() {
+        // No arg: report current persona instead of switching.
+        let map = sessions.lock().await;
+        let current = map
+            .get(&chat_id)
+            .and_then(|s| s.active_persona.clone())
+            .unwrap_or_else(|| "ctrl (default)".to_string());
+        drop(map);
+        bot.send_message(
+            chat_id,
+            format!(
+                "Active persona: {current}\n\nUse /switch <name> to change. \
+                 Available: ctrl, izzie, cto (or cto-assistant)"
+            ),
+        )
+        .await?;
+        return Ok(());
+    }
+    // Canonicalize aliases to match REPL's /switch behavior
+    // (src/repl/agent_commands.rs): `cto`/`cto assistant` -> `cto-assistant`,
+    // and `default` -> `ctrl`. Pass through any other agent TOML stems
+    // unchanged so user-authored personas keep working.
+    let stem = match arg.to_lowercase().trim() {
+        "ctrl" | "default" => "ctrl".to_string(),
+        "izzie" => "izzie".to_string(),
+        "cto" | "cto-assistant" | "cto assistant" => "cto-assistant".to_string(),
+        _ => arg.clone(),
+    };
+    // `/switch ctrl` / `/switch default` clears the persona, restoring
+    // the default ctrl/PM runner used before any switch.
+    if stem == "ctrl" {
+        let mut map = sessions.lock().await;
+        let entry = map
+            .entry(chat_id)
+            .or_insert_with(|| ChatSession::new((**project_path).clone()));
+        entry.active_persona = None;
+        drop(map);
+        bot.send_message(chat_id, "✓ Switched back to ctrl (default)")
+            .await?;
+        return Ok(());
+    }
+    // Validate the persona resolves. Mirror ctrl's resolution order:
+    // project-local agents dir first, then ~/.open-mpm/agents/.
+    let session_project_path = {
+        let map = sessions.lock().await;
+        map.get(&chat_id)
+            .map(|s| s.project_path.clone())
+            .unwrap_or_else(|| (**project_path).clone())
+    };
+    let persona_path = session_project_path
+        .join(".open-mpm")
+        .join("agents")
+        .join(format!("{stem}.toml"));
+    if persona_path.exists() || home_persona_exists(&stem) {
+        let mut map = sessions.lock().await;
+        let entry = map
+            .entry(chat_id)
+            .or_insert_with(|| ChatSession::new((**project_path).clone()));
+        entry.active_persona = Some(stem.clone());
+        drop(map);
+        bot.send_message(chat_id, format!("✓ Switched to {arg}"))
+            .await?;
+    } else {
+        bot.send_message(
+            chat_id,
+            format!("Unknown persona: {arg}. Available: ctrl, izzie, cto (or cto-assistant)"),
+        )
+        .await?;
+    }
+    Ok(())
+}
